PyCharmJiJing/src/api/main.py
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
import logging
from prometheus_client import Counter, Histogram

app = FastAPI(title="基金预测系统")

# 性能指标
PREDICTION_REQUEST_COUNT = Counter(
    'prediction_requests_total',
    'Total number of prediction requests'
)
PREDICTION_LATENCY = Histogram(
    'prediction_latency_seconds',
    'Time spent processing prediction requests'
)

class PredictionRequest(BaseModel):
    fund_code: str
    prediction_days: int

@app.post("/predict")
async def predict_fund(request: PredictionRequest):
    try:
        PREDICTION_REQUEST_COUNT.inc()
        with PREDICTION_LATENCY.time():
            # 实现预测逻辑
            prediction = {"predicted_value": 0.0}  # 示例返回值
            return prediction
    except Exception as e:
        logging.error(f"预测错误: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e)) 