PyCharmJiJing/src/models/fund_predictor.py
import torch
import torch.nn as nn
from typing import Dict

class LSTMPredictor(nn.Module):
    def __init__(self, input_dim: int, hidden_dim: int, num_layers: int):
        super().__init__()
        self.lstm = nn.LSTM(
            input_dim, 
            hidden_dim,
            num_layers,
            batch_first=True
        )
        self.fc = nn.Linear(hidden_dim, 1)
        
    def forward(self, x):
        lstm_out, _ = self.lstm(x)
        predictions = self.fc(lstm_out[:, -1, :])
        return predictions

class ModelTrainer:
    def __init__(self, config: Dict):
        self.config = config
        self.model = self._build_model()
        
    def _build_model(self) -> nn.Module:
        return LSTMPredictor(
            input_dim=self.config['input_dim'],
            hidden_dim=self.config['hidden_dim'],
            num_layers=self.config['num_layers']
        )
        
    def train(self, train_data, valid_data):
        """
        实现模型训练逻辑
        """
        pass 