PyCharmJiJing/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
pytorch>=1.9.0
fastapi>=0.68.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
mlflow>=1.20.0
prometheus-client>=0.11.0
pytest>=6.2.0 